run command in active terminal, more refactor
