+++ src/extension.ts
@@ -12,6 +12,7 @@ import resetFn from "./functions/reset";
 import copyCommandFn from "./functions/copyCommand";
 import Command from "./models/command";
 import { ExecCommands } from "./models/exec_commands";
+import runCommandInActiveTerminalFn from "./functions/runCommandInActiveTerminal";
 
 const terminals: Array<vscode.Terminal> = [];
 
@@ -67,9 +68,13 @@ export function activate(context: vscode.ExtensionContext) {
           terminal.sendText(c[i].command);
           terminal.show();
         } else {
-          throw Error("Unable to find the command in state");
+          throw new ReadableError("Unable to find the command in state");
         }
       } catch (e) {
+        if (e instanceof ReadableError) {
+          vscode.window.showErrorMessage(e.message);
+          return;
+        }
         vscode.window.showErrorMessage("Unable to execute the command");
       }
     }
@@ -77,27 +82,7 @@ export function activate(context: vscode.ExtensionContext) {
 
   const runCommandInActiveTerminal = vscode.commands.registerCommand(
     ExecCommands.runCommandInActiveTerminal,
-    (item: TreeItem) => {
-      let c: any;
-      try {
-        if (item.contextValue === "child-workspace") {
-          c = Command.getWorkspaceCommands(context);
-        } else if (item.contextValue === "child-global") {
-          c = Command.getGlobalCommands(context);
-        }
-        const i = c.findIndex((d: any) => d.id === item.cmdId);
-        if (i > -1) {
-          const terminalId = `${c[i].name}-${utils.generateString(5)}`;
-          const terminal = vscode.window.createTerminal(terminalId);
-          terminal.sendText(c[i].command);
-          terminal.show();
-        } else {
-          throw Error("Unable to find the command in state");
-        }
-      } catch (e) {
-        vscode.window.showErrorMessage("Unable to execute the command");
-      }
-    }
+    runCommandInActiveTerminalFn(context)
   );
   const reset = vscode.commands.registerCommand(
     ExecCommands.reset,
