+++ src/models/error.ts
@@ -1,0 +1,6 @@
+class ReadableError extends Error {
+  constructor(message: string) {
+    super(message);
+    this.name = "ReadableError";
+  }
+}
